aboba

+++ bot/handlers/checker.py
@@ -1,5 +1,6 @@
 from aiogram import Router
 from aiogram.types import Message
+from aiogram.enums.chat_type import ChatType
 
 from bot.bot import bot
 
@@ -7,4 +8,5 @@ router = Router()
 
 @router.message()
 async def checker(message: Message):
-    await bot.delete_message(chat_id=message.chat.id,message_id=message.message_id)
+    if message.chat.type != ChatType.GROUP and message.chat.type != ChatType.SUPERGROUP:
+        await bot.delete_message(chat_id=message.chat.id,message_id=message.message_id)
